+++ forui/CHANGELOG.md
@@ -13,7 +13,6 @@
 
 
 ### `FSelect`
-
 * Fix `FSelect` flickering when selecting an item on desktop.
 * Fix `FSelect` not scrolling to selected item when opened if the item is really far down the list.
 * Fix `FSelect` layout shifting when scrolling through a long list of items.
@@ -30,7 +29,9 @@
 
 
 ### Others
+* **Breaking** Change `FPersistentSheetController.shown` to `FPersistentSheetController.status`.
 * **Breaking** Change `FPopoverController.shown` to `FPopoverController.status`.
+* **Breaking** Change `FTooltipController.shown` to `FTooltipController.status`.
 
 
 ## 0.13.1